pyproject.toml
[project]
name = "uom-term1-ai-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "hf-xet>=1.2.0",
    "huggingface-hub[hf-xet]>=0.36.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "numpy>=2.3.4",
    "pydantic>=2.12.4",
    "readability-lxml>=0.8.4.1",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
